WowCar/requirements.txt
asgiref==3.5.2
Django==4.0.4
django-environ==0.8.1
django-filter==22.1
djangorestframework==3.14.0
install==1.3.5
psycopg2-binary==2.9.3
pytz==2022.4
sqlparse==0.4.2
